Feat: Updated Multiple Features
Updated CurrentInventory class to reflect changes to database: Added fields for loader, workOrder, loadTime, and loadNum.

Added method to update database with scanned material to reflect which load and workOrder it was on.

Added method for manual removal of material from current load / reception.

Increased functionality of input boxes by adding usage of Model-View-ViewModel design. User Input is now saved between pages and is used to calculate number of items remaining to be loaded.

Added function to display number of bundles remaining in load and take user directly to confirmation screen when requested amount is reached.