Fix Issue with waiting for matching BL's before loading next screen
Fixed issue where correct screen / dialog was not being displayed after searching by base heat. Added methods in view model to make this a synchronous call because the UI is dependent on the result f the room query. Will be implementing further logic to handle the different cases which will appear from usage.