Create Methods for Browser Automation
Created method waitForEitherElement for cases when user is directed to one of 2 distinct pages and action must be take based on which one is supplied.

Added methods to parent classes Automated Session and Reception to better maintain separation of concerns and increase maintainability of application.

Added method logic for logging into TC3.

Added method logic for creating a new reception in TC3.