Add Outline of Back-end Logic for Creating Rusal and Boscus Shipments
Added Shipment class and ShipmentWithRelease Class to ensure that child classes always implement the same logic resulting in a project which is easier to maintain in the long term.

Created controller method for shipment with endpoint at /api/browser/shipment to receive data from front-end. Note will need to add username and password fields to front-end form.

Moved getClerkInitials method from AlgomaReception to SeleniumHelper to it can be utilized in additional classes.

Added skeleton classes for Bosucs and Rusal Shipments and Releases to outline logic to be implemented for each.