Refactor: Remove Unnecessary ViewModels
Started refactor of application via removing unnecessary viewmodels for splashscreen and mainmenu. Moved relevant logic to mainactivity viewmodel as it be needed again later in the activity and should not require calls to repository when unecessary.

Created enum SessionType to ensure that the only types of session are General, Shipment, and Reception. This is done to increase the maintainability  and readability of code.