Refactor: ReturnedBundle Screen and ViewModel
Refactored Returned enum to track the different returnable values of items from database. and ensure that all possible returnable options are accounted for.

Replaced "General" Main Menu Button destination with ToBeImplemented Screen to ensure application is in a fully functional state and users cannot access unimplemented features.

Ensure that Scanned Heats are formatted the same as manually entered heat numbers when reaching Returned Bundle ViewModel.

Added withContext to suspended repository operations to ensure that operations are always called on Dispatchers.IO thread.

Updated Main Menu Screen to warn users who are about to create a new session of a different type when there are currently items added to a session of a different type.

Renamed pieceCount to quantity in RusalItem to bring back in-line with data received from the server.

Separated Update methods in repository such that receptions and shipments have unique logic and methods.