Add Functionality to Sync Reception Progress In Background
Added functionality to sync reception progress in background every without user intervention. Updated such that for every 20 items added to the reception, the first 10 scanned are uploaded to the server.

Updated review screen such that the list of items shown only shows up to the latest 10 added items. Changed as users do not require the ability to see more than the latest 10 items in the field. As such displaying more would only make it more difficult to find the item they are searching for.